fix(repos): rename Go and C++ http_archive names to match go-jsonnet
`rules_jsonnet` uses the names `jsonnet_go` and `cpp_jsonnet` for the
different jsonnet implementations, but `jsonnet_go` in turn uses the name
`jsonnet` for the repository that `rules_jsonnet` calls `cpp_jsonnet`.

Building works just fine, but when trying bazel query on anything that
depends on `rules_jsonnet` it fails as it can't find the repo `jsonnet`.

https://github.com/google/go-jsonnet/blob/cd59751527e0e79994aec71a8c7f954ece3da4af/bazel/repositories.bzl#L17-L38
https://github.com/bazelbuild/rules_jsonnet/blob/11039058a409b3c6a11c6548bb75871b4b295c53/jsonnet/jsonnet.bzl#L800-L816